feat: Add apix_open_unix/tcp_client

+++ src/apix-posix.c
@@ -31,9 +31,9 @@ struct posix_sink {
     int nfds;
 };
 
-// unix domain socket server
-
-static struct posix_sink __unix_s_sink;
+/*
+ * unix domain socket server
+ */
 
 static int unix_s_open(struct apisink *sink, const char *addr)
 {
@@ -80,7 +80,8 @@ static int unix_s_close(struct apisink *sink, int fd)
     if (sinkfd == NULL)
         return -1;
     close(sinkfd->fd);
-    unlink(sinkfd->addr);
+    if (strcmp(sink->id, APISINK_UNIX_S) == 0)
+        unlink(sinkfd->addr);
     sinkfd_destroy(sinkfd);
     return 0;
 }
@@ -167,9 +168,117 @@ static apisink_ops_t unix_s_ops = {
     .poll = unix_s_poll,
 };
 
-// tcp server
+/*
+ * unix domain socket client
+ */
 
-static struct posix_sink __tcp_s_sink;
+static int unix_c_open(struct apisink *sink, const char *addr)
+{
+    int fd = socket(PF_UNIX, SOCK_STREAM, 0);
+    if (fd == -1)
+        return -1;
+
+    int rc = 0;
+    struct sockaddr_un sockaddr = {0};
+    sockaddr.sun_family = PF_UNIX;
+    snprintf(sockaddr.sun_path, sizeof(sockaddr.sun_path), "%s", addr);
+
+    rc = connect(fd, (struct sockaddr *)&sockaddr, sizeof(sockaddr));
+    if (rc == -1) {
+        close(fd);
+        return -1;
+    }
+
+    struct sinkfd *sinkfd = sinkfd_new();
+    sinkfd->fd = fd;
+    snprintf(sinkfd->addr, sizeof(sinkfd->addr), "%s", addr);
+    sinkfd->sink = sink;
+    list_add(&sinkfd->node_sink, &sink->sinkfds);
+    list_add(&sinkfd->node_ctx, &sink->ctx->sinkfds);
+
+    struct posix_sink *unix_c_sink = container_of(sink, struct posix_sink, sink);
+    FD_SET(fd, &unix_c_sink->fds);
+    unix_c_sink->nfds = fd + 1;
+
+    return fd;
+}
+
+static int unix_c_close(struct apisink *sink, int fd)
+{
+    struct sinkfd *sinkfd = find_sinkfd_in_apisink(sink, fd);
+    if (sinkfd == NULL)
+        return -1;
+    close(sinkfd->fd);
+    sinkfd_destroy(sinkfd);
+    return 0;
+}
+
+static int unix_c_send(struct apisink *sink, int fd, const void *buf, size_t len)
+{
+    UNUSED(sink);
+    return send(fd, buf, len, 0);
+}
+
+static int unix_c_recv(struct apisink *sink, int fd, void *buf, size_t size)
+{
+    UNUSED(sink);
+    return recv(fd, buf, size, 0);
+}
+
+static int unix_c_poll(struct apisink *sink)
+{
+    struct posix_sink *unix_c_sink = container_of(sink, struct posix_sink, sink);
+
+    struct timeval tv = { 0, 0 };
+    fd_set recvfds;
+    memcpy(&recvfds, &unix_c_sink->fds, sizeof(recvfds));
+
+    int nr_recv_fds = select(unix_c_sink->nfds, &recvfds, NULL, NULL, &tv);
+    if (nr_recv_fds == -1) {
+        if (errno == EINTR)
+            return 0;
+        LOG_ERROR("[select] (%d) %s", errno, strerror(errno));
+        return -1;
+    }
+
+    struct sinkfd *pos, *n;
+    list_for_each_entry_safe(pos, n, &sink->sinkfds, node_sink) {
+        if (nr_recv_fds == 0) break;
+
+        if (!FD_ISSET(pos->fd, &recvfds))
+            continue;
+
+        nr_recv_fds--;
+
+        int nread = recv(pos->fd, atbuf_write_pos(pos->rxbuf),
+                            atbuf_spare(pos->rxbuf), 0);
+        if (nread == -1) {
+            LOG_DEBUG("[recv] (%d) %s", errno, strerror(errno));
+            sinkfd_destroy(pos);
+        } else if (nread == 0) {
+            LOG_DEBUG("[recv] (%d) finished");
+            sinkfd_destroy(pos);
+        } else {
+            atbuf_write_advance(pos->rxbuf, nread);
+            gettimeofday(&pos->ts_poll_recv, NULL);
+        }
+    }
+
+    return 0;
+}
+
+static apisink_ops_t unix_c_ops = {
+    .open = unix_c_open,
+    .close = unix_c_close,
+    .ioctl = NULL,
+    .send = unix_c_send,
+    .recv = unix_c_recv,
+    .poll = unix_c_poll,
+};
+
+/*
+ * tcp server
+ */
 
 static int tcp_s_open(struct apisink *sink, const char *addr)
 {
@@ -219,28 +328,73 @@ static int tcp_s_open(struct apisink *sink, const char *addr)
     return fd;
 }
 
-static int tcp_s_close(struct apisink *sink, int fd)
-{
-    struct sinkfd *sinkfd = find_sinkfd_in_apisink(sink, fd);
-    if (sinkfd == NULL)
-        return -1;
-    close(sinkfd->fd);
-    sinkfd_destroy(sinkfd);
-    return 0;
-}
-
 static apisink_ops_t tcp_s_ops = {
     .open = tcp_s_open,
-    .close = tcp_s_close,
+    .close = unix_s_close,
     .ioctl = NULL,
     .send = unix_s_send,
     .recv = unix_s_recv,
     .poll = unix_s_poll,
 };
 
-// serial
+/*
+ * tcp client
+ */
 
-static struct posix_sink __serial_sink;
+static int tcp_c_open(struct apisink *sink, const char *addr)
+{
+    int fd = socket(PF_INET, SOCK_STREAM, 0);
+    if (fd == -1)
+        return -1;
+
+    uint32_t host;
+    uint16_t port;
+    char *tmp = strdup(addr);
+    char *colon = strchr(tmp, ':');
+    *colon = 0;
+    host = inet_addr(tmp);
+    port = htons(atoi(colon + 1));
+    free(tmp);
+
+    int rc = 0;
+    struct sockaddr_in sockaddr = {0};
+    sockaddr.sin_family = PF_INET;
+    sockaddr.sin_addr.s_addr = host;
+    sockaddr.sin_port = port;
+
+    rc = connect(fd, (struct sockaddr *)&sockaddr, sizeof(sockaddr));
+    if (rc == -1) {
+        close(fd);
+        return -1;
+    }
+
+    struct sinkfd *sinkfd = sinkfd_new();
+    sinkfd->fd = fd;
+    sinkfd->listen = 1;
+    snprintf(sinkfd->addr, sizeof(sinkfd->addr), "%s", addr);
+    sinkfd->sink = sink;
+    list_add(&sinkfd->node_sink, &sink->sinkfds);
+    list_add(&sinkfd->node_ctx, &sink->ctx->sinkfds);
+
+    struct posix_sink *tcp_c_sink = container_of(sink, struct posix_sink, sink);
+    FD_SET(fd, &tcp_c_sink->fds);
+    tcp_c_sink->nfds = fd + 1;
+
+    return fd;
+}
+
+static apisink_ops_t tcp_c_ops = {
+    .open = tcp_c_open,
+    .close = unix_c_close,
+    .ioctl = NULL,
+    .send = unix_c_send,
+    .recv = unix_c_recv,
+    .poll = unix_c_poll,
+};
+
+/*
+ * serial
+ */
 
 static int serial_open(struct apisink *sink, const char *addr)
 {
@@ -396,28 +550,83 @@ static apisink_ops_t serial_ops = {
 
 int apix_enable_posix(struct apix *ctx)
 {
-    apisink_init(&__unix_s_sink.sink, APISINK_UNIX_S, unix_s_ops);
-    apix_add_sink(ctx, &__unix_s_sink.sink);
+    // unix_s
+    struct posix_sink *unix_s_sink = calloc(1, sizeof(struct posix_sink));
+    apisink_init(&unix_s_sink->sink, APISINK_UNIX_S, unix_s_ops);
+    apix_add_sink(ctx, &unix_s_sink->sink);
 
-    apisink_init(&__tcp_s_sink.sink, APISINK_TCP_S, tcp_s_ops);
-    apix_add_sink(ctx, &__tcp_s_sink.sink);
+    // unix_c
+    struct posix_sink *unix_c_sink = calloc(1, sizeof(struct posix_sink));
+    apisink_init(&unix_c_sink->sink, APISINK_UNIX_C, unix_c_ops);
+    apix_add_sink(ctx, &unix_c_sink->sink);
 
-    apisink_init(&__serial_sink.sink, APISINK_SERIAL, serial_ops);
-    apix_add_sink(ctx, &__serial_sink.sink);
+    // tcp_s
+    struct posix_sink *tcp_s_sink = calloc(1, sizeof(struct posix_sink));
+    apisink_init(&tcp_s_sink->sink, APISINK_TCP_S, tcp_s_ops);
+    apix_add_sink(ctx, &tcp_s_sink->sink);
+
+    // tcp_c
+    struct posix_sink *tcp_c_sink = calloc(1, sizeof(struct posix_sink));
+    apisink_init(&tcp_c_sink->sink, APISINK_TCP_C, tcp_c_ops);
+    apix_add_sink(ctx, &tcp_c_sink->sink);
+
+    // serial
+    struct posix_sink *serial_sink = calloc(1, sizeof(struct posix_sink));
+    apisink_init(&serial_sink->sink, APISINK_SERIAL, serial_ops);
+    apix_add_sink(ctx, &serial_sink->sink);
 
     return 0;
 }
 
 void apix_disable_posix(struct apix *ctx)
 {
-    apix_del_sink(ctx, &__unix_s_sink.sink);
-    apisink_fini(&__unix_s_sink.sink);
+    struct apisink *pos, *n;
+    list_for_each_entry_safe(pos, n, &ctx->sinks, node) {
+        // unix_s
+        if (strcmp(pos->id, APISINK_UNIX_S) == 0) {
+            struct posix_sink *unix_s_sink =
+                container_of(pos, struct posix_sink, sink);
+            apix_del_sink(ctx, &unix_s_sink->sink);
+            apisink_fini(&unix_s_sink->sink);
+            free(unix_s_sink);
+        }
 
-    apix_del_sink(ctx, &__tcp_s_sink.sink);
-    apisink_fini(&__tcp_s_sink.sink);
+        // unix_c
+        if (strcmp(pos->id, APISINK_UNIX_C) == 0) {
+            struct posix_sink *unix_c_sink =
+                container_of(pos, struct posix_sink, sink);
+            apix_del_sink(ctx, &unix_c_sink->sink);
+            apisink_fini(&unix_c_sink->sink);
+            free(unix_c_sink);
+        }
 
-    apix_del_sink(ctx, &__serial_sink.sink);
-    apisink_fini(&__serial_sink.sink);
+        // tcp_s
+        if (strcmp(pos->id, APISINK_TCP_S) == 0) {
+            struct posix_sink *tcp_s_sink =
+                container_of(pos, struct posix_sink, sink);
+            apix_del_sink(ctx, &tcp_s_sink->sink);
+            apisink_fini(&tcp_s_sink->sink);
+            free(tcp_s_sink);
+        }
+
+        // tcp_c
+        if (strcmp(pos->id, APISINK_TCP_C) == 0) {
+            struct posix_sink *tcp_c_sink =
+                container_of(pos, struct posix_sink, sink);
+            apix_del_sink(ctx, &tcp_c_sink->sink);
+            apisink_fini(&tcp_c_sink->sink);
+            free(tcp_c_sink);
+        }
+
+        // serial
+        if (strcmp(pos->id, APISINK_SERIAL) == 0) {
+            struct posix_sink *serial_sink =
+                container_of(pos, struct posix_sink, sink);
+            apix_del_sink(ctx, &serial_sink->sink);
+            apisink_fini(&serial_sink->sink);
+            free(serial_sink);
+        }
+    }
 }
 
 #endif
